refactor: extract worker and watch modules
