+++ babelarr/__init__.py
@@ -1,14 +1,17 @@
 """Babelarr package."""
 
-from .app import Application, SrtHandler
+from .app import Application
 from .config import Config
 from .queue_db import QueueRepository
 from .translator import LibreTranslateClient, Translator
+from .watch import SrtHandler
+from .worker import TranslationTask
 
 __all__ = [
     "Config",
     "Application",
     "SrtHandler",
+    "TranslationTask",
     "QueueRepository",
     "Translator",
     "LibreTranslateClient",
